rename directory and e2e flow to be Discover Screen now since it isn't a sheet anymore
