+++ e2e/discoverSheetFlow.spec.js
@@ -5,7 +5,7 @@ import * as Helpers from './helpers';
 const ios = device.getPlatform() === 'ios';
 const android = device.getPlatform() === 'android';
 
-describe('Discover Sheet Flow', () => {
+describe('Discover Screen Flow', () => {
   it('Should show the welcome screen', async () => {
     await Helpers.checkIfVisible('welcome-screen');
   });
