+++ src/components/discover-screen/DPICard.tsx
