+++ src/components/discover-screen/EdgeFade.js
