+++ src/components/discover-screen/PulseIndexSection.js
