+++ src/components/discover-screen/androidCustomComponents/customBackground.tsx
