+++ src/components/exchange/ExchangeSearch.tsx
@@ -18,7 +18,7 @@ import Animated, {
   withTiming,
 } from 'react-native-reanimated';
 import Spinner from '../../assets/chartSpinner.png';
-import DiscoverSheetContext from '../discover-sheet/DiscoverSheetContext';
+import DiscoverSheetContext from '../discover-screen/DiscoverScreenContext';
 import { ClearInputDecorator, Input } from '../inputs';
 import { analytics } from '@/analytics';
 import { ImgixImage } from '@/components/images';
